requirements.txt
# Base dependencies
pip>=23.0
setuptools>=65.5.0
wheel>=0.40.0

# Framework Web
fastapi==0.95.2
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.0.0
websockets>=12.0

# Base de données et Cache
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
aioredis>=2.0.0

# Traitement Audio
pydub==0.25.1
librosa==0.10.1
numpy>=1.20.0,<1.25.0
soundfile==0.12.1
pyacoustid==1.2.2
chromaprint==0.5
ffmpeg-python>=0.2.0
av>=10.0.0

# Détection Musicale
musicbrainzngs==0.7.1
requests==2.31.0
aiohttp==3.9.1

# Machine Learning
tensorflow>=2.12.0

# Génération de Rapports
pandas<2.0.0
openpyxl==3.1.2
fpdf==1.7.2
XlsxWriter==3.1.9

# Utilitaires
python-dotenv==1.0.0
pydantic==1.10.8
typing-extensions>=4.5.0
psutil==5.9.6
python-json-logger==2.0.7
tenacity==8.2.3

# Tests
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-redis==3.1.2
pytest-benchmark==4.0.0
pytest-xdist==3.5.0
httpx==0.25.2
coverage==7.3.2
aioresponses==0.7.4

# Linting et Qualité de Code
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0

# Monitoring et Logging
prometheus-client==0.19.0
structlog==23.2.0
